--- about.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>О проекте</title>
-</head>
-<body>
-    <h1>О проекте</h1>
-    <ul>
-        <li>
-            <a href="/">
-                Главная
-            </a>
-        </li>
-        <li>
-            <a href="/authors/">
-                Авторы
-            </a>
-        </li>
-        <li>
-            <a href="/contacts/">
-                Контакты
-            </a>
-        </li>
-        <li>
-            <a href="/about/">
-                О проекте
-            </a>
-        </li>
-    </ul>
-
-    <h1>Параметры</h1>
-    <ul>
-        {% for param in params.items() %}
-            <li>{{param}}</li>
-        {% endfor %}
-    </ul>
-
-</body>
-</html>
