--- authors.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Authors</title>
-</head>
-<body>
-    <h1>Авторы</h1>
-    <ul>
-        <li>
-            <a href="/">
-                Главная
-            </a>
-        </li>
-        <li>
-            <a href="/authors/">
-                Авторы
-            </a>
-        </li>
-        <li>
-            <a href="/contacts/">
-                Контакты
-            </a>
-        </li>
-        <li>
-            <a href="/about/">
-                О проекте
-            </a>
-        </li>
-    </ul>
-    <ul>
-        {% for author in object_list %}
-            <li>{{author.name}}</li>
-        {% endfor %}
-    </ul>
-
-    <h1>Параметры</h1>
-    <ul>
-        {% for param in params.items() %}
-            <li>{{param}}</li>
-        {% endfor %}
-    </ul>
-
-</body>
-</html>
