+++ framework_wsgi/wsgi.py
@@ -1,79 +1,3 @@
-import json
-import datetime
-
-from templator import render
-
-
-def contacts_view(params, method):
-    def save_user_message(params):
-        now = datetime.datetime.now()
-        dt_str = now.strftime("%d-%m-%Y %H:%M:%S")
-        message_dict = '{0}\r\n'.format(json.dumps(params))
-        with open("user_messages.txt", "a") as text_file:
-            text_file.write("{0}::Message from user: {1}".format(dt_str, message_dict))
-        return
-
-    if method == 'GET':
-        output = render('contacts.html', params=params)
-    elif method == 'POST':
-        save_user_message(params)
-        output = render('contacts.html', params=params, result_message='Сообщение успешно отправлено')
-    else:
-        return '405 ERROR', [b'Unsupport method']
-
-    b_output = bytes(output, encoding='utf-8')
-    return '200 OK', [b_output]
-
-
-def index_view(params, method):
-    output = render('main.html', params=params)
-    b_output = bytes(output, encoding='utf-8')
-    return '200 OK', [b_output]
-
-
-def authors_view(params, method):
-    output = render('authors.html', object_list=[{'name': 'Leo'}, {'name': 'Kate'}], params=params)
-    b_output = bytes(output, encoding='utf-8')
-    return '200 OK', [b_output]
-
-
-def about_view(params, method):
-    output = render('about.html', params=params)
-    b_output = bytes(output, encoding='utf-8')
-    return '200 OK', [b_output]
-
-
-def not_found_404_view(params, method):
-    return '404 WHAT', [b'404 PAGE Not Found']
-
-
-class Other:
-    def __call__(self, params):
-        print(params)
-        return '200 OK', [b'<h1>other</h1>']
-
-
-routes = {
-    '/': index_view,
-    '/authors/': authors_view,
-    '/about/': about_view,
-    '/contacts/': contacts_view,
-    '/other/': Other(),
-    '': not_found_404_view
-}
-
-
-# Front controllers
-def secret_front(request):
-    request['secret'] = 'some secret'
-
-
-def other_front(request):
-    request['key'] = 'key'
-
-
-fronts = [secret_front, other_front]
-
 
 class Application:
     def __init__(self, routes, fronts):
@@ -82,6 +6,9 @@ class Application:
 
     def __call__(self, environ, start_response):
         path = environ['PATH_INFO']
+        # Если в конце нет '/' то добавляем
+        if path[-1] != '/':
+            path = path + '/'
         if path in self.routes:
             view = self.routes[path]
         else:
@@ -138,4 +65,4 @@ class Application:
         return params
 
 
-application = Application(routes, fronts)
+#application = Application(routes, fronts)
