+++ templates/authors.html
@@ -1,0 +1,17 @@
+{% extends 'base.html' %}
+
+    {% block content %}
+    <ul>
+        {% for author in context.object_list %}
+            <li>{{author.name}}</li>
+        {% endfor %}
+    </ul>
+
+    <h1>Параметры</h1>
+    <ul>
+        {% for param in context.params.items() %}
+            <li>{{param}}</li>
+        {% endfor %}
+    </ul>
+
+    {% endblock %}
