+++ templates/base.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>{{ context.title }}</title>
+</head>
+<body>
+    <h1>{{ context.title }}</h1>
+
+    {% include './includes/main_menu.html' %}
+    {% block content %}
+    {% endblock %}
+</body>
+</html>
