+++ templates/includes/main_menu.html
@@ -1,0 +1,6 @@
+    <ul>
+        <li><a href="/">Главная</a></li>
+        <li><a href="/authors/">Авторы</a></li>
+        <li><a href="/contacts/">Контакты</a></li>
+        <li><a href="/about/">О проекте</a></li>
+    </ul>
