+++ urls.py
@@ -1,0 +1,10 @@
+from views import *
+
+urls = {
+    '/': index_view,
+    '/authors/': authors_view,
+    '/about/': about_view,
+    '/contacts/': contacts_view,
+    '/other/': Other(),
+    '': not_found_404_view
+}
