+++ user_messages.txt
@@ -1,2 +1,7 @@
-30-08-2020 15:39:14::Message from user: {"fname": "John", "lname": "Doe", "email": "example%40domain.ru", "message": "111111111111"}
-30-08-2020 15:39:18::Message from user: {"fname": "John", "lname": "Doe", "email": "example%40domain.ru", "message": "22222222222222"}
+30-08-2020 15:57:25::Message from user: {"fname": "John", "lname": "Doe", "email": "example%40domain.ru", "message": "%D0%B2%D0%B0%D0%BF"}
+30-08-2020 15:59:09::Message from user: {"fname": "John", "lname": "Doe", "email": "example%40domain.ru", "message": "%D0%BF%D0%B0%D0%B2"}
+01-09-2020 00:20:08::Message from user: {"fname": "John", "lname": "Doe", "email": "example%40domain.ru", "message": "ertert"}
+01-09-2020 00:20:09::Message from user: {"fname": "John", "lname": "Doe", "email": "example%40domain.ru", "message": ""}
+01-09-2020 00:20:28::Message from user: {"fname": "John", "lname": "Doe", "email": "example%40domain.ru", "message": "wef"}
+18-09-2020 18:10:31::Message from user: {"fname": "John", "lname": "Doe", "email": "example%40domain.ru", "message": "dfbdfb"}
+18-09-2020 18:13:09::Message from user: {"fname": "John", "lname": "Doe", "email": "example%40domain.ru", "message": "ergerg"}
